refactor: 실거래가 조회 API 리팩토링 및 불필요한 코드 제거 -  유즈케이스 동작명 수정 - infrastructure 밑 파일 명 수정 - axios 싱글톤 방식으로 수정 - xml 파싱 utils로 분리 - 서비스 키 불러오는 위치 변경 - http 상태 코드 추가 - route 응답, 요청에 dto 사용 및 코드 리팩토링
