+++ app/api/real-estate/search/transaction-price/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { RealEstateTransactionUseCase } from '../../../../../backend/realEstate/applications/usecases/RealEstateTransactionUseCase';
+import { GetRealEstateTransactionUseCase } from '../../../../../backend/realEstate/applications/usecases/GetRealEstateTransactionUseCase';
 import { GetRealEstateTransactionRequest } from '../../../../../backend/realEstate/applications/dtos/GetRealEstateTransactionRequest';
+import { GetRealEstateTransactionResponse } from '../../../../../backend/realEstate/applications/dtos/GetRealEstateTransactionResponse';
 
 /**
  * 실거래가 조회 API
@@ -8,9 +9,8 @@ import { GetRealEstateTransactionRequest } from '../../../../../backend/realEsta
  */
 export async function GET(request: NextRequest) {
   try {
+    // 1. URL에서 쿼리 파라미터를 가져와서
     const { searchParams } = new URL(request.url);
-    
-    // 쿼리 파라미터 추출
     const LAWD_CD = searchParams.get('LAWD_CD');
     const DEAL_YMD = searchParams.get('DEAL_YMD');
     const numOfRows = searchParams.get('numOfRows') || '10';
@@ -31,38 +31,27 @@ export async function GET(request: NextRequest) {
       );
     }
     
-    // 환경변수에서 서비스키 가져오기
-    const serviceKey = process.env.RTMSDATA_TRANSACTION_PRICE_KEY;
-    if (!serviceKey) {
-      return NextResponse.json(
-        { success: false, message: '서비스키가 설정되지 않았습니다.' },
-        { status: 500 }
-      );
-    }
-    
-    // 요청 데이터 구성
+    // 2. 쿼리를 위해 전달할 DTO 생성한 후
     const requestData: GetRealEstateTransactionRequest = {
       LAWD_CD,
       DEAL_YMD,
-      serviceKey,
       numOfRows,
       pageNo,
     };
     
-    // 유스케이스 실행
-    const useCase = new RealEstateTransactionUseCase();
-    const response = await useCase.findAll(requestData);
+    // 3. 쿼리를 위해 전달할 DTO를 Usecase에게 전달해서 실행한다.
+    const useCase = new GetRealEstateTransactionUseCase();
+    const response: GetRealEstateTransactionResponse = await useCase.execute(requestData);
     
-    // 성공 응답
+    // 응답 데이터를 JSON 형식으로 반환 (DTO 타입 명시)
     return NextResponse.json({
       success: true,
-      data: response,
-    });
+      data: response as GetRealEstateTransactionResponse,
+    }, { status: 200 });
     
   } catch (error) {
     console.error('실거래가 조회 API 에러:', error);
     
-    // 에러 응답
     return NextResponse.json(
       { 
         success: false,
